+++ packages/input-autocomplete/src/types.ts
@@ -6,6 +6,8 @@ import type {
     BottomSheetSelectMobileProps,
 } from '@alfalab/core-components-select/shared';
 
+type OnInputTypeReason = 'close' | 'change';
+
 export interface InputAutocompleteCommonProps
     extends Omit<
         BaseSelectProps,
@@ -39,7 +41,7 @@ export interface InputAutocompleteCommonProps
     /**
      * Обработчик ввода
      */
-    onInput?: (value: string) => void;
+    onInput?: (value: string, reason?: OnInputTypeReason) => void;
 }
 
 type MobileProps = {