+++ packages/pure-cell/src/components/addon/component.tsx
@@ -29,6 +29,16 @@ type Props = {
      * Используется модификатор -addon
      */
     dataTestId?: string;
+
+    /**
+     * Клик по контенту аддона.
+     */
+    onClick?: () => void;
+};
+
+const ADDON_COMPONENT: Record<string, keyof Pick<React.ReactHTML, 'button' | 'section'>> = {
+    button: 'button',
+    section: 'section',
 };
 
 export const Addon: React.FC<Props> = ({
@@ -36,15 +46,37 @@ export const Addon: React.FC<Props> = ({
     verticalAlign = 'top',
     addonPadding = 'default',
     dataTestId,
+    onClick,
 }) => {
     const pureCellContext = useContext(PureCellContext);
 
+    const Component = onClick ? ADDON_COMPONENT.button : ADDON_COMPONENT.section;
+
+    const onMouseEvents = {
+        onClick: (e: React.MouseEvent<HTMLButtonElement>) => {
+            e.stopPropagation();
+            onClick?.();
+        },
+        onMouseEnter: () => {
+            pureCellContext.unsetMainHover?.();
+        },
+        onMouseLeave: () => {
+            pureCellContext.setMainHover?.();
+        },
+        onMouseDown: (e: React.MouseEvent<HTMLButtonElement>) => {
+            e.stopPropagation();
+        },
+    };
+
     return (
-        <section
-            className={cn(styles.component, styles[addonPadding], styles[verticalAlign])}
+        <Component
+            className={cn(styles.component, styles[addonPadding], styles[verticalAlign], {
+                [styles.button]: onClick,
+            })}
             data-test-id={getDataTestId(dataTestId || pureCellContext.dataTestId, 'addon')}
+            {...(onClick && onMouseEvents)}
         >
             {children}
-        </section>
+        </Component>
     );
 };