+++ packages/date-range-input/CHANGELOG.md
@@ -1,5 +1,14 @@
 # @alfalab/core-components-date-range-input
 
+## 3.3.11
+
+### Patch Changes
+
+-   Обновлены зависимости
+    -   calendar@7.6.0
+    -   input@14.4.0
+    -   icon-button@6.5.6
+
 ## 3.3.10
 
 ### Patch Changes