+++ packages/input-autocomplete/CHANGELOG.md
@@ -1,5 +1,21 @@
 # Change Log
 
+## 12.1.0
+
+### Minor Changes
+
+<sup><time>27.04.2024</time></sup>
+
+### [#1171](https://github.com/core-ds/core-components/pull/1171)
+
+-   Добавлен параметр reason для onInput события
+
+### Patch Changes
+
+-   Обновлены зависимости
+    -   input@14.4.0
+    -   select@17.4.1
+
 ## 12.0.6
 
 ### Patch Changes