+++ packages/modal/CHANGELOG.md
@@ -1,5 +1,18 @@
 # Change Log
 
+## 9.2.3
+
+### Patch Changes
+
+<sup><time>27.04.2024</time></sup>
+
+### [#1181](https://github.com/core-ds/core-components/pull/1181)
+
+-   В компоненте Modal.Header теперь учитывается как onClose пропса из Modal, так и из Modal.Header
+
+-   Обновлены зависимости
+    -   navigation-bar@0.8.7
+
 ## 9.2.2
 
 ### Patch Changes