+++ packages/toast/CHANGELOG.md
@@ -1,5 +1,12 @@
 # Change Log
 
+## 5.2.12
+
+### Patch Changes
+
+-   Обновлены зависимости
+    -   toast-plate@6.3.7
+
 ## 5.2.11
 
 ### Patch Changes